requirements.txt
python-libnmap>=0.7.3
requests>=2.32.3
rich>=13.9.4
Jinja2>=3.1.6
packaging>=25.0
pytest>=8.0
pytest-mock>=3.14
responses>=0.25
